--- Cyan-Stars/Assets/Scripts/ChartEditor/NewCorrectAudioBar.cs
@@ -1,9 +1,0 @@
-using CyanStars.ChartEditor;
-using UnityEngine;
-
-public class NewCorrectAudioBar : MonoBehaviour, ILoopScrollRectItem
-{
-    public void ScrollCellIndex(int index)
-    {
-    }
-}
