--- Cyan-Stars/Assets/Scripts/ChartEditor/NewCorrectAudioBar.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 5465889a3d32f8b49b3bca3a1c6ab989
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
